add mypy, remove outdated manual json tests, update test json
