+++ Pipfile
@@ -12,6 +12,7 @@ flake8 = "==7.0.0"
 httpx = "==0.26.0"
 black = "==24.3.0"
 pytest-cov = "==5.0.0"
+mypy = "==1.9.0"
 
 [packages]
 fastapi = "==0.109.1"
